// site/www/src/components/stream/diagrams/types.ts
// Copyright 2026 Synnax Labs, Inc.
//
// Use of this software is governed by the Business Source License included in the file
// licenses/BSL.txt.
//
// As of the Change Date specified in that file, in accordance with the Business Source
// License, use of this software will be governed by the Apache License, Version 2.0,
// included in the file licenses/APL.txt.

import type { CSSProperties } from "react";

export type NodeIcon = React.ComponentType<{ style?: CSSProperties }>;

export interface NodeDef {
  id: string;
  x: number;
  y: number;
  w: number;
  h: number;
  label: string;
  icon: NodeIcon;
}

export interface EdgeDef {
  from: string;
  to: string;
}

export type DiagramVariant = "schematic" | "pill";

export interface DiagramDef {
  viewBox: string;
  nodes: NodeDef[];
  edges: EdgeDef[];
  variant?: DiagramVariant;
}